:truck: #553 - refactor: alter directory structure
