+++ frontend/src/pages/destructionlist/detail/pages/edit/DestructionListEditPage.stories.tsx
