+++ frontend/src/pages/destructionlist/detail/pages/edit/index.ts
