+++ frontend/src/pages/destructionlist/detail/pages/process-review/components/index.ts
