+++ frontend/src/pages/destructionlist/detail/pages/process-review/index.ts
